install bear/fastly-module
